surfaceshader
n the code above, the lightningShader function takes three parameters: the output parameter representing the connection to the shader output, the eyeColor parameter representing the color of the eye, and the lightColor parameter representing the color of the lightning effect. The function creates a SurfaceShader node with the specified name, adds the eyeColor and lightColor attributes, and connects them to the provided values. Finally, it connects the output of the shader to the specified output parameter.

To use the lightningShader function, you can pass a PyNode object representing the output parameter and tuples representing the eye color and light color. In the example, pm.PyNode('lightningShader_output') is used to get the output parameter node by name, and (0.1, 0.1, 0.1) and (1.0, 1.0, 1.0) are used as the eye color and light color, respectively.

Keep in mind that this is a basic example, and you may need to adjust the shader attributes and connections based on your specific requirements and the RenderMan shader network you are building in Maya.